+++ sshd-core/src/main/java/org/apache/sshd/client/kex/DHGClient.java
@@ -269,11 +269,15 @@ public class DHGClient extends AbstractDHClientKeyExchange {
                     "KeyExchange CA signature verification failed for key type=" + keyAlg + " of key ID=" + keyId);
         }
 
+        // OpenSSH < 10.3:
+        //
         // "As a special case, a zero-length "valid principals" field means the certificate is valid for
         // any principal of the specified type."
         // See https://github.com/openssh/openssh-portable/blob/master/PROTOCOL.certkeys
         //
         // Empty principals in a host certificate mean the certificate is valid for any host.
+        //
+        // OpenSSH >= 10.3: such certificates are never valid.
         Collection<String> principals = openSshKey.getPrincipals();
         if (!GenericUtils.isEmpty(principals)) {
             /*
@@ -296,6 +300,9 @@ public class DHGClient extends AbstractDHClientKeyExchange {
                 throw new SshException(SshConstants.SSH2_DISCONNECT_KEY_EXCHANGE_FAILED,
                         "KeyExchange signature verification failed, could not determine connect host for key ID=" + keyId);
             }
+        } else if (!CoreModuleProperties.ALLOW_EMPTY_CERTIFICATE_PRINCIPALS.getRequired(session)) {
+            throw new SshException(SshConstants.SSH2_DISCONNECT_KEY_EXCHANGE_FAILED,
+                    "KeyExchange signature verification failed because the certificate has no principals; key ID=" + keyId);
         }
 
         if (!GenericUtils.isEmpty(openSshKey.getCriticalOptions())) {