+++ sshd-core/src/main/java/org/apache/sshd/core/CoreModuleProperties.java
@@ -799,6 +799,12 @@ public final class CoreModuleProperties {
         }
     });
 
+    /**
+     * Whether to allow SSH user or host certificates without principals. OpenSSH < 10.3 considered such certificates to
+     * be valid for any principal; since OpenSSH 10.3 such certificates are rejected.
+     */
+    public static final Property<Boolean> ALLOW_EMPTY_CERTIFICATE_PRINCIPALS = Property.bool("empty-cert-principals", false);
+
     private CoreModuleProperties() {
         throw new UnsupportedOperationException("No instance");
     }